Implement null consensus
This provides a scaffolding for an alternative implementation of
IValidatorGroup. The sample implementation provided in the commit (null
consensus) is a debug tool that allows to test all the consensus infra
independently of a real consensus algorithm.

Null consensus is capable of finalizing blocks when all nodes are
trusted and network is reasonably reliable.